+++ docs/conf.py
@@ -16,8 +16,8 @@ sys.path.insert(0, str(Path(__file__).parent.parent))
 project = 'ColorCast'
 copyright = '2024, ColorCast Contributors'
 author = 'ColorCast Contributors'
-release = '2.2.0'
-version = '2.2.0'
+release = '2.3.0'
+version = '2.3.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration